Working on intgrating push notifcation and WEB Push API with the backend
