--- src/main/java/com/revature/StudyForce/StudyForceApplication.java
@@ -1,13 +1,0 @@
-package com.revature.StudyForce;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StudyForceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(StudyForceApplication.class, args);
-	}
-
-}
