--- src/main/java/com/revature/StudyForce/flashcard/model/Difficulty.java
@@ -1,11 +1,0 @@
-package com.revature.StudyForce.flashcard.model;
-
-/**
- * Enum used to represent the difficulty of a flashcard
- * @author Edson Rodriguez
- */
-public enum Difficulty {
-    EASY,
-    MEDIUM,
-    HARD
-}
