+++ src/main/java/com/revature/studyforce/flashcard/model/Difficulty.java
@@ -1,0 +1,11 @@
+package com.revature.studyforce.flashcard.model;
+
+/**
+ * Enum used to represent the difficulty of a flashcard
+ * @author Edson Rodriguez
+ */
+public enum Difficulty {
+    EASY,
+    MEDIUM,
+    HARD
+}
