+++ src/main/java/com/revature/studyforce/flashcard/model/Flashcard.java
@@ -1,0 +1,43 @@
+package com.revature.studyforce.flashcard.model;
+
+import com.sun.istack.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+import com.revature.studyforce.user.model.User;
+
+import java.sql.Timestamp;
+import java.time.LocalDateTime;
+
+/**
+ * Flashcard model
+ * @author Luke
+ */
+
+@Entity
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Flashcard {
+
+    @Id
+    @GeneratedValue
+    private int id;
+    @NotNull
+    @ManyToOne
+    private User creator;
+    //TODO: waiting on Topic model
+//    @NotNull
+//    @ManyToOne
+//    private Topic topic;
+    @Column(nullable = false)
+    private String question;
+    private int questionDifficultyTotal = 0;
+    private int questionDifficultyAverage = 0;
+    private Timestamp createdTime = Timestamp.valueOf(LocalDateTime.now());
+    private Timestamp resolutionTime = Timestamp.valueOf(LocalDateTime.now());
+
+}
