+++ src/main/java/com/revature/studyforce/flashcard/model/Rating.java
@@ -1,0 +1,36 @@
+package com.revature.studyforce.flashcard.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import com.revature.studyforce.user.model.User;
+
+import javax.persistence.*;
+import javax.validation.constraints.NotNull;
+
+/**
+ * Model used to store the difficulty ratings of a flashcards
+ *
+ *@author Edson Rodriguez
+ */
+@Entity
+@Table(name = "timesheets")
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Rating {
+    @Id
+    @GeneratedValue
+    private int id;
+    @NotNull
+    @OneToOne
+    private Flashcard flashcardId;
+    @NotNull
+    @OneToOne
+    private User user;
+    @NotNull
+    @Column(name = "rating")
+    private Difficulty ratingValue;
+
+}
