+++ src/main/java/com/revature/studyforce/flashcard/repository/RatingRepository.java
@@ -1,0 +1,16 @@
+package com.revature.studyforce.flashcard.repository;
+
+import com.revature.studyforce.flashcard.model.Rating;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+/**
+ * Represents the Repository for the Rating model
+ *
+ * @author Edson Rodriguez
+ */
+@Repository
+public interface RatingRepository extends JpaRepository<Rating,Integer> {
+
+}
