+++ src/main/java/com/revature/studyforce/notifcation/model/StackTrace.java
@@ -1,0 +1,29 @@
+package com.revature.studyforce.notifcation.model;
+
+
+import com.revature.studyforce.user.model.User;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@Table(name = "stackTrace")
+public class StackTrace {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "stackTraceId")
+    private int stackTraceId;
+
+    @Column(name = "userId")
+    private int userId;
+
+
+
+
+}
