+++ src/main/java/com/revature/studyforce/notifcation/model/UserSubscriptionID.java
@@ -1,0 +1,17 @@
+package com.revature.studyforce.notifcation.model;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Embeddable;
+import java.io.Serializable;
+@Embeddable
+@NoArgsConstructor
+@Data
+public class UserSubscriptionID implements Serializable {
+
+    private int user;
+
+    private int subscription;
+
+}
