+++ src/main/java/com/revature/studyforce/notifcation/repo/SubscriptionRepository.java
@@ -1,0 +1,11 @@
+package com.revature.studyforce.notifcation.repo;
+
+import com.revature.studyforce.notifcation.model.Subscription;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SubscriptionRepository extends JpaRepository<Subscription , Integer> {
+
+
+}
