+++ src/main/resources/application.properties
@@ -1,8 +1,9 @@
 spring.datasource.username=sa
 spring.datasource.password=
-spring.datasource.url=jdbc:h2:~/test
+spring.datasource.url=jdbc:h2:tcp://localhost/~/test
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
+spring.jpa.database=h2
 
 spring.jpa.hibernate.ddl-auto=update
